refactor(library): decouple SAM3OpenVINO from the torch SAM3 class
Loading a pre-exported OpenVINO IR required importing the torch-backed
SAM3 model definition. Torch is needed to build and export the model,
not to reuse the artifact afterwards.

SAM3OpenVINO imported SAM3 for exactly one thing: SAM3.card(), which
just returns the constant _SAM3_CARD from the already torch-free
_card module. The three other imported symbols (SAM3_MODEL_ID,
CanvasConfig, Sam3PromptMode) are a str, a dataclass and an Enum that
merely lived in the torch module.

Move those three into the existing dependency-free constants module -
the module whose docstring already stated that purpose - and read the
card from _card directly. sam3.py re-imports them so existing import
paths keep working.
